refactor(studio): relocate analysis/cloud/video out of modules; delete src/modules (Tauri-shell step 2c)
The last three modules aren't pure facades: video/analysis hold the binary job
coordinators (JobMap + worker thread + the `video://progress` / `analysis://progress`
events) and cloud holds the keychain-backed object-store adapter — real Tauri-side
glue that can't be eliminated, only relocated. Move each to a top-level subsystem
folder (`src/analysis/`, `src/cloud/`, `src/video/`, like `src/ai/` and
`src/runtime/`), rewrite paths (`crate::modules::<f>::` → `crate::<f>::`), move the
`pub mod` decls to lib.rs, and DELETE `src/modules/` entirely.

`src/` is now the Tauri shell: `commands/` (IPC handlers) + `composition.rs` +
lib.rs/main.rs, plus per-subsystem glue folders and the infra files
(inference.rs/gpu.rs/store.rs/schema.rs) pending crate extraction. No `domain/` or
`modules/` folder remains. No IPC behavior change. Green: default `cargo check`.
