+++ apps/studio/src-tauri/src/analysis/mod.rs
