+++ apps/studio/src-tauri/src/analysis/service.rs
