+++ apps/studio/src-tauri/src/cloud/model.rs
@@ -1,7 +1,7 @@
 //! Re-export shim.
 //!
 //! The cloud DTOs now live in `vailabel_cloud::contracts`. This shim keeps the
-//! historical path `crate::modules::cloud::model::*` valid so `commands.rs`
+//! historical path `crate::cloud::model::*` valid so `commands.rs`
 //! compiles unchanged. Reverting the cloud migration = restoring this file (and
 //! `service.rs`) to their prior bodies.
