--- apps/studio/src-tauri/src/modules/mod.rs
@@ -1,3 +1,0 @@
-pub mod analysis;
-pub mod cloud;
-pub mod video;
