+++ apps/studio/src-tauri/src/video/model.rs
@@ -2,7 +2,7 @@
 //!
 //! The video annotation data model and request payloads now live in the
 //! `vailabel-video` crate (`domain` / `contracts`). This shim keeps
-//! `crate::modules::video::model::*` valid for the video service, ffmpeg, and
+//! `crate::video::model::*` valid for the video service, ffmpeg, and
 //! interpolation code that remain in the binary. `now_iso` in `VideoJob::new`
 //! now resolves `vailabel_shared::now_iso` (same implementation).
